Fix RFC 5155 NSEC3 test by using `-P` instead of `-p`, i.e. ACTUALLY apply opt-out.
Split out some of the ldns-signzone `-b` behaviours into their own flags for `dnst signzone`.

Add a new `-L` flag which dumps a list of owner name to hash mappings as comments before the zone output proper (similar to the mappings shown in the RFC 4035 and RFC 5155 signed zone examples).

Use the new flags to produce much closer output to that of the example signed zone in RFC 5155 making it easier to see if we match its behaviour or not (we do).