Package as cascade-dnst instead of dnst. (#131)
To avoid Cascade installing dnst and uninstalling ldns-utils that Cascade users may want to use such as ldns-verify-zone, which dnst doesn't provide yet. Also because this branch of dnst depends on an as yet unreleased version of the domain crate.

Update Cargo.lock bcause cargo-deb complains that it needs updating.

Co-authored-by: Jannik Peters <jannik@nlnetlabs.nl>